Add Image Transfer resources
Motivation
OKD on oVirt is using terraform to stand up a cluster, by creating
masters, all from a specific qcow url. This url changes with the version
and needs to uploaded per the cluster version. The intention is to make
it as seemless as possible for users that install OKD on oVirt, without
a pre-step of creating that specific template.

Modification
1. Add an ovirt_image resource, that will handle disk uploads from urls, or
local files (which counts as urls).
oVirt api exposes 1 or 2 valid urls to upload the image to: either a
proxy url, when the proxy is the only accesible node for the client, and
the host url. The provider will first try to reach the host_url, which
is quicker, and if its not available it will use the proxy url.
Update on and image transfer is not supported.
2. Add ovirt_template resource that will create a template from a VM
and together with image_transfer it is now possible to upload an image
and make it a template, with interimediate step of creating a VM from
the new uploaded disk.

Result
Given this main.tf file, a disk with the name rhcos will be created
under the provided storage domain id and a template will be made of it:

```
resource "ovirt_image_transfer" "rhcos_transfer" {
  alias             = "rhcos"
  source_url        = "https://path/to/image"
  storage_domain_id = "d787bf6b-fae1-4a3e-b773-2ac466599d29"
  sparse            = true
}

resource "ovirt_vm" "tmpvm" {
  name = "tmpvm-for-${ovirt_image_transfer.rhcos_transfer.alias}"
  cluster_id = "5c8f6906-f14b-43ee-83df-5f800f36eb70"
  block_device {
    disk_id   = ovirt_image_transfer.rhcos_transfer.disk_id
    interface = "virtio_scsi"
  }
}

resource "ovirt_template" "rhcos_template" {
  name = "${ovirt_image_transfer.rhcos_transfer.alias}"
  cluster_id = ovirt_vm.tmpvm.cluster_id
  // create from vm
  vm_id = ovirt_vm.tmpvm.id
}

```

Signed-off-by: Roy Golan <rgolan@redhat.com>