Override disk attachment on VM resource create
Motivation
New VMs inherit their bootable disk attachement details from the template. A RHEL/RHCOS, CentOS, Ubuntu or any os would likely want to override those on creation. We want a bigger disk for some VMs, different driver, and do that in one-shot definition, without needing to go and create disk attachment resources and then terraform-refresh them.

Modification
The existing 'block_device' element in vm resource can take 'disk_id' optionally instead of mandatory. When the `disk_id` is lacking, the bootable disk that is attached to this newly created vm is updated with the details in `block_device`.

Result
We have a template with OS disk of 16Gib.
We want to create a new VM and extend the OS disk to 120Gib:
```hcl
resource "ovirt_vm" "masters" {
  ...
  template_id = xyz  // the template with 16GiB disk
  block_device {
      size = 120     // this will extend the disk to 120GiB
  }
}
```