ovirt_tag: if tag exists continue in resourceOvirtTagCreate
If user try to use in terraform twice or more the statement:
resource "ovirt_tag" "cluster_tag" will raise error that
the rag already exists not allowing others vms to be tagged.
This patch detects if the tag exists and set the id from
the tag to the function continue without problem.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1806471
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1841380

Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>