Remove vm disk by default
When we try to remove a VM with terraform unless that VM was create from a template then the VM disk is not removed from the infrastructure.

This raises 2 problems:
1. It is inconsistent with the oVirt API, when you remove a VM with the oVirt API then unless a flag is specified then the VM disk will be removed.
2. Terraform should leave a clean infrastructure when removing resources, without any "extra" unmanaged resources.

This patch changes the behavior of ovirt vm removal, now when you remove a VM the underlying disk is also removed.

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>