improve wait for transfer to complete
Currently ovirt_image_transfer waits for the disk to become OK to
determine the transfer is complete, however, the disk's status change to
OK prematurely in ovirt.

This patch aims to implement a more reliable check by bringing the
correlation ID into the mix and making sure all commands and their
children are completed before moving on to checking the state of the
entities.

Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>