introduce auto pinning policy
This patch will introduce the AutoPinningPolicy. This will allow VMs to
automatically get the configuration for CPUs and NUMA. For high
performance VMs, it will be automatically set to `existing` policy.

- `adjust` policy will change the CPU topology based on the host
  topology, will set the vCPU pinning and NUMA nodes.
- `existing` policy will set the vCPU pinning and NUMA nodes with the
  existing VM CPU topology.
- `disabled` won't do nothing, and can used to disable the default high
  performance VM policy value.

Each VM using a policy will be pinned to each host within the cluster
with allowed migration. Once the user will add a host, he needs to
update the VMs, adding the new host to the pinned hosts.

This patch bump the go-ovirt SDK.

Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>