remove vm_list from affinity group resource
You can now add a VM to an affinity group on creation, a feature we didn't have when the affinity group resource was implemented.

Because a VM can join an affinity group in 2 ways it could lead to inconsistencies in the tfstate.

We are removing the ability to collect existing VMs into an affinity group since this path is the least recommended approch because it will cause the affinity group to be best effort