remove existing for HP VMs
This patch will remove the undesired behavior of setting auto pinning
policy to existing when the VM is set with High Performance profile and
not provided with any auto pinning policy input.

The existing policy is going to be blocked in ovirt, we should prevent
users from hitting an error while creating a VM without specifying a
value that might cause it.